requirements.txt
Flask==2.3.3 
Flask-SQLAlchemy==3.0.5 
Flask-WTF==1.2.1 
Werkzeug==3.0.1 
python-dotenv==1.0.1 
Flask-Migrate==4.0.7 
scikit-learn==1.5.2 
numpy==1.26.4